+++ src/backends/sentry_backend_crashpad.cpp
@@ -597,14 +597,22 @@ crashpad_backend_startup(
     }
 #endif
 
+    crashpad::CrashpadInfo *crashpad_info
+        = crashpad::CrashpadInfo::GetCrashpadInfo();
+
     if (!options->system_crash_reporter_enabled) {
         // Disable the system crash reporter. Especially on macOS, it takes
         // substantial time *after* crashpad has done its job.
-        crashpad::CrashpadInfo *crashpad_info
-            = crashpad::CrashpadInfo::GetCrashpadInfo();
         crashpad_info->set_system_crash_reporter_forwarding(
             crashpad::TriState::kDisabled);
     }
+
+    if (options->crashpad_limit_stack_capture_to_sp) {
+        // Enable stack capture limit to work around Wine/Proton TEB issues
+        crashpad_info->set_limit_stack_capture_to_sp(
+            crashpad::TriState::kEnabled);
+    }
+
     return 0;
 }