+++ src/sentry_core.c
@@ -314,6 +314,9 @@ sentry_flush(uint64_t timeout)
 {
     int rv = 0;
     SENTRY_WITH_OPTIONS (options) {
+        if (options->enable_logs) {
+            sentry__logs_force_flush();
+        }
         rv = sentry__transport_flush(options->transport, timeout);
     }
     return rv;
@@ -454,6 +457,16 @@ sentry_user_consent_get(void)
             (long *)&options->user_consent);
     }
     return rv;
+}
+
+int
+sentry_user_consent_is_required(void)
+{
+    int required = 0;
+    SENTRY_WITH_OPTIONS (options) {
+        required = options->require_user_consent;
+    }
+    return required;
 }
 
 void
@@ -925,6 +938,39 @@ sentry_remove_extra_n(const char *key, size_t key_len)
 {
     SENTRY_WITH_SCOPE_MUT (scope) {
         sentry_value_remove_by_key_n(scope->extra, key, key_len);
+    }
+}
+
+void
+sentry_set_attribute(const char *key, sentry_value_t attribute)
+{
+    SENTRY_WITH_SCOPE_MUT (scope) {
+        sentry__scope_set_attribute(scope, key, attribute);
+    }
+}
+
+void
+sentry_set_attribute_n(
+    const char *key, size_t key_len, sentry_value_t attribute)
+{
+    SENTRY_WITH_SCOPE_MUT (scope) {
+        sentry__scope_set_attribute_n(scope, key, key_len, attribute);
+    }
+}
+
+void
+sentry_remove_attribute(const char *key)
+{
+    SENTRY_WITH_SCOPE_MUT (scope) {
+        sentry__scope_remove_attribute(scope, key);
+    }
+}
+
+void
+sentry_remove_attribute_n(const char *key, size_t key_len)
+{
+    SENTRY_WITH_SCOPE_MUT (scope) {
+        sentry__scope_remove_attribute_n(scope, key, key_len);
     }
 }