+++ src/sentry_options.c
@@ -52,6 +52,7 @@ sentry_options_new(void)
     opts->crashpad_wait_for_upload = false;
     opts->enable_logging_when_crashed = true;
     opts->propagate_traceparent = false;
+    opts->crashpad_limit_stack_capture_to_sp = false;
     opts->symbolize_stacktraces =
     // AIX doesn't have reliable debug IDs for server-side symbolication,
     // and the diversity of Android makes it infeasible to have access to debug
@@ -489,6 +490,13 @@ sentry_options_set_crashpad_wait_for_upload(
 }
 
 void
+sentry_options_set_crashpad_limit_stack_capture_to_sp(
+    sentry_options_t *opts, int enabled)
+{
+    opts->crashpad_limit_stack_capture_to_sp = !!enabled;
+}
+
+void
 sentry_options_set_shutdown_timeout(
     sentry_options_t *opts, uint64_t shutdown_timeout)
 {
@@ -734,6 +742,19 @@ int
 sentry_options_get_enable_logs(const sentry_options_t *opts)
 {
     return opts->enable_logs;
+}
+
+void
+sentry_options_set_logs_with_attributes(
+    sentry_options_t *opts, int logs_with_attributes)
+{
+    opts->logs_with_attributes = !!logs_with_attributes;
+}
+
+int
+sentry_options_get_logs_with_attributes(const sentry_options_t *opts)
+{
+    return opts->logs_with_attributes;
 }
 
 #ifdef SENTRY_PLATFORM_LINUX